+++ src/test/java/com/moaplace/service/MailSendServiceTest.java
@@ -1,15 +1,10 @@
 package com.moaplace.service;
 
-import java.util.List;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-import com.moaplace.vo.MemberVO;
 
 import lombok.extern.log4j.Log4j;
 
@@ -24,12 +19,6 @@ public class MailSendServiceTest {
 	@Autowired
 	private MailSendService service;
 	
-	@Autowired
-	private MemberService memberService;
-	
-	@Autowired
-	private PasswordEncoder passwordEncoder;
-	
 	@Test
 	public void test() {
 		
@@ -42,16 +31,4 @@ public class MailSendServiceTest {
 		}
 	}
 	
-	@Test
-	public void selectListTest() {
-		List<MemberVO> vo = memberService.selectAll();
-		for( MemberVO m : vo) {
-			System.out.println(m);
-		}
-	}
-	
-	@Test
-	public void passwordEncoderTest() {
-		log.info(passwordEncoder.encode("1234"));
-	}
 }
